Change the cursor from union to struct.
It is said to be a bad idea to use union in C++, so I changed it to a plain
=struct=. Then it turns out that previously the cursor drawn on the screen is a
member of the editor class, instead of a member of the buffer class. So I shall
correct this mistake later, and implement some methods for the cursor to move:
it now only has constructors and private data members and we cannot change it
conveniently.